+++ packages/app/src/index.ts
@@ -1,5 +1,18 @@
 export * from './app'
 export * from './request'
+import type { Request } from './request'
 export * from './response'
-export * from './router'
+import type { Response } from './response'
 export * from './extend'
+export type { ErrorHandler } from './onError'
+import { NextFunction, Handler as RHandler, AsyncHandler as RAsyncHandler, SyncHandler as RSyncHandler } from '@tinyhttp/router'
+
+export type Handler = RHandler<Request, Response>
+
+export type AsyncHandler = RAsyncHandler<Request, Response>
+
+export type SyncHandler = RSyncHandler<Request, Response>
+
+export type { NextFunction }
+
+export type { Request, Response }