+++ packages/cookie/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/cookie
 
+## 0.1.19
+
+### Patch Changes
+
+- Split app into req, res, send and router
+
 ## 0.1.18
 
 ### Patch Changes