+++ packages/dotenv/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/dotenv
 
+## 0.0.9
+
+### Patch Changes
+
+- Split app into req, res, send and router
+
 ## 0.0.8
 
 ### Patch Changes