+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinyhttp/markdown
 
+## 0.1.54
+
+### Patch Changes
+
+- Split app into req, res, send and router
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.73
+
 ## 0.1.53
 
 ### Patch Changes