+++ packages/ping/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/ping
 
+## 0.0.3
+
+### Patch Changes
+
+- Split app into req, res, send and router
+
 ## 0.0.2
 
 ### Patch Changes