+++ packages/session/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @tinyhttp/session
 
+## 0.0.4
+
+### Patch Changes
+
+- Split app into req, res, send and router
+- Updated dependencies [undefined]
+  - @tinyhttp/cookie@0.1.19
+  - @tinyhttp/cookie-signature@0.1.16
+
 ## 0.0.3
 
 ### Patch Changes